修改了购物车与商品表间的关联

+++ SmartBuy/src/main/java/com/hit/sb/cart/model/CartModel.java
@@ -12,7 +12,7 @@ public class CartModel implements Serializable {
 		private int uid;
 		private int pid;
 		private int num;
-		private double price;
+		private float price;
 		private Date createdtime;
 		private Date modifiedtime;
 	  
@@ -43,7 +43,7 @@ public class CartModel implements Serializable {
 	public double getPrice() {
 		return price;
 	}
-	public void setPrice(double d) {
+	public void setPrice(float d) {
 		this.price = d;
 	}
 	public Date getCreatedtime() {
